refactor(guild): extract IUnitOfWork as the commit boundary and document kick/ban target ordering
