+++ services/guild/tests/Guild.UnitTests/Application/PutOverwriteHandlerTests.cs
@@ -82,7 +82,6 @@ public sealed class PutOverwriteHandlerTests
 		Assert.Single(overwrites.Store);
 		Assert.Equal(OverwriteTargetType.Role, overwrites.Store.Values.Single().TargetType);
 		Assert.Equal(1, overwrites.AddCount);
-		Assert.Equal(1, overwrites.SaveChangesCount);
 	}
 
 	[Fact]
@@ -101,7 +100,6 @@ public sealed class PutOverwriteHandlerTests
 		Assert.Equal(8L, only.Deny);
 		Assert.Equal(1, overwrites.AddCount);
 		Assert.Equal(1, overwrites.UpdateCount);
-		Assert.Equal(2, overwrites.SaveChangesCount);
 	}
 
 	[Fact]
