+++ services/guild/tests/Guild.UnitTests/Application/UpdateCategoryHandlerTests.cs
@@ -155,7 +155,6 @@ public sealed class UpdateCategoryHandlerTests
 		Assert.Equal("New", result.Value.Name);
 		Assert.Equal(12, result.Value.Position);
 		Assert.Equal(1, catRepo.UpdateCount);
-		Assert.Equal(1, catRepo.SaveChangesCount);
 	}
 
 	[Fact]
