+++ services/guild/tests/Guild.UnitTests/Application/UpdateChannelHandlerTests.cs
@@ -68,7 +68,6 @@ public sealed class UpdateChannelHandlerTests
 		Assert.Equal("new", result.Value.Name);
 		Assert.Equal(4, result.Value.Position);
 		Assert.Equal(1, channels.UpdateCount);
-		Assert.Equal(1, channels.SaveChangesCount);
 	}
 
 	[Fact]
