+++ services/guild/tests/Guild.UnitTests/Fakes/FakeChannelPermissionOverwriteRepository.cs
@@ -12,7 +12,6 @@ internal sealed class FakeChannelPermissionOverwriteRepository : IChannelPermiss
 	public int AddCount { get; private set; }
 	public int UpdateCount { get; private set; }
 	public int RemoveCount { get; private set; }
-	public int SaveChangesCount { get; private set; }
 
 	public Task<IReadOnlyList<ChannelPermissionOverwrite>> GetForChannelAsync(
 		long channelId,
@@ -64,12 +63,6 @@ internal sealed class FakeChannelPermissionOverwriteRepository : IChannelPermiss
 	{
 		_store.Remove(overwrite.Id);
 		RemoveCount++;
-	}
-
-	public Task SaveChangesAsync(CancellationToken cancellationToken = default)
-	{
-		SaveChangesCount++;
-		return Task.CompletedTask;
 	}
 
 	internal void Seed(ChannelPermissionOverwrite overwrite) => _store[overwrite.Id] = overwrite;
