+++ services/guild/tests/Guild.UnitTests/Fakes/FakeChannelRepository.cs
@@ -12,7 +12,6 @@ internal sealed class FakeChannelRepository : IChannelRepository
 	public int AddCount { get; private set; }
 	public int UpdateCount { get; private set; }
 	public int RemoveCount { get; private set; }
-	public int SaveChangesCount { get; private set; }
 
 	public Task<Channel?> GetByIdAsync(long channelId, CancellationToken cancellationToken = default)
 	{
@@ -59,12 +58,6 @@ internal sealed class FakeChannelRepository : IChannelRepository
 	{
 		_store.Remove(channel.Id);
 		RemoveCount++;
-	}
-
-	public Task SaveChangesAsync(CancellationToken cancellationToken = default)
-	{
-		SaveChangesCount++;
-		return Task.CompletedTask;
 	}
 
 	internal void Seed(Channel channel) => _store[channel.Id] = channel;
